changed folder structure and fixed pep8 and linting
