+++ api/.gitignore
@@ -1,3 +1,1 @@
 vars.bat
-__pycache__/
-app/__pycache__/
