+++ api/run.py
@@ -1,7 +1,7 @@
 """Run this file to start the application"""
 import os
 
-from app.routes.route import create_app
+from app.app import create_app
 
 
 config_name = os.environ.get('FLASK_ENV')
