+++ frontend/css/answer.css
@@ -8,7 +8,6 @@ body {
     
 }
 
-
 .header {
     padding: 100px;
     width: 100%;
@@ -16,15 +15,12 @@ body {
   
 }
 
-
-
 .header h1 {
     font-size: 60px;
     color: green;
     margin-left: 490px;
     margin-top: 1px;
     font-family:  Georgia, serif;
-
 
 }
